### 2016-10-14 #80 @cboettig
review request for EML
Parse and serialize Ecological Metadata Language ('EML') files into S4 objects.

```
Package: EML
Title: Read and Write Ecological Metadata Language Files
Version: 1.0.0.0
    person("Carl", "Boettiger", role = c("aut", "cre"),
           email = "cboettig@gmail.com"),
    person("Matt", "Jones", role="aut"),
    person("Bryce", "Mecum", role = "ctb", email = "brycemecum@gmail.com"),
    person("Maëlle", "Salmon", role = "ctb", email = "maelle.salmon@yahoo.se")
    )
Maintainer: Carl Boettiger <cboettig@gmail.com>
Description: Parse and serialize Ecological Metadata Language ('EML') files into
    S4 objects.
License: FreeBSD
    XML,
    tools
Collate:
    'classes.R'
    'classes-stmml.R'
    'coercions.R'
    'eml_find.R'
    'get_unitList.R'
    'get_attributes.R'
    'eml_get.R'
    'eml_validate.R'
    'get_TextType.R'
    'get_coverage.R'
    'is_standardUnit.R'
    'literature_coercions.R'
    'methods.R'
    'read_eml.R'
    'set_TextType.R'
    'set_attributes.R'
    'set_coverage.R'
    'set_methods.R'
    'set_physical.R'
    'set_unitList.R'
    'write_eml.R'
    'xml-s4.R'
    'zzz.R'
```

https://github.com/ropensci/EML

Researchers publishing data who want to use a formal, machine-readable metadata standard to describe their data files, and/or researchers consuming data published with EML metadata (e.g. [NEON](http://www.neonscience.org/) or [LTER](https://lternet.edu/) data.)

Not that I'm aware of.

  - (_Do not submit your package separately to JOSS_)
- [Leah Wasser](https://github.com/lwasser)
- [Peter Slaughter](https://github.com/gothub)
- [Bryce Mecum](https://github.com/amoeba) (also a minor contributor to the package)
- [Corinna Gries](https://github.com/cgries)
- [Chris Jones](https://github.com/csjx)
