### 2016-10-14 ropensci/EML#181 @cboettig
Prepare CRAN release
Hi @mbjones,

 I think I'm ready to release an initial version of EML to CRAN.  I've opened an onboarding request on ropensci to get some more eyes on it before I send it off, https://github.com/ropensci/onboarding/issues/80.  There's still plenty of higher-level functionality that could be improved, but I think the low level stuff is stable and since you're team is using it in production, it's about time for a CRAN release.  Please weigh in here if there's anything else you'd like to see addressed before a first release to CRAN.

Thanks!
