### 2016-10-18 #80 @noamross
Second reviewer: @cgries
