### 2016-11-01 #80 @sckott
@gmbecker @cgries - hey there, it's been 17 days, please get your review in by Nov 04, thanks :smiley_cat: (ropensci-bot)
