### 2016-11-10 #80 @noamross
Thank you for the review @gmbecker!

Regarding the full package import of XML, I agree that it's fine.  There are a few points in our review criteria where we need to clarify requirement v. recommendation, and how we draw the line.  I will make a note of this one for our next update.
