### 2016-11-10 ropensci/EML#183 @cboettig
Address issues raised in rOpenSci Onboarding review
@mbjones , others watching this repo,

We've got the first of our very helpful reviews from the ropensci onboarding process: https://github.com/ropensci/onboarding/issues/80.  Many good suggestions there which I will try and open in individual issues here when we get a chance.  Meanwhile just wanted to get feedback from anyone using EML currently about which issues raised in the review (as well as outstanding issues here, see [CRAN release milestone](https://github.com/ropensci/EML/milestone/7)) really ought to be addressed prior to a CRAN release + zenodo archive.  My intuition is that if EML is being used for stuff as is, that it might be better to have stable release perhaps before all these issues are addressed, but definitely worth thinking about those that most impact the API.

Thanks for any thoughts! Also we should get another review still coming.
