### 2016-11-11 ropensci/EML#183 @amoeba
All of the issues raised in that review seem valid and like they would improve the package.

However, the only one that I really felt strongly about was the Constructors comment. I seem to have enough success using `new` everywhere in my code, but I appreciate that the user experience would be greatly improved with constructor functions. The use of `new` instead of constructors has been a point of confusion when I've helped others try to use the package.

*That said...*

The package is working great right now, as is. It's made a huge, positive impact on the types of metadata authoring we do. My vote goes for pushing a CRAN release with the current API and pushing towards having constructors at a later date. I think your intuition is spot on.

As an aside, we're using the EML package via [helper functions](https://github.com/NCEAS/arcticdatautils/blob/master/R/eml.R) in an internal R package https://github.com/NCEAS/arcticdatautils. The helper functions are optimized for our specific use but I thought I'd share.
