### 2016-11-11 ropensci/EML#183 @mbjones
I'm with @amoeba on his comments on release with constructors.  There are a lot of good suggestions in the review though, and some of them might not delay a cran release. For example, we could patch up the eml_get return types, fix some of the autogenerated code issues, and fix the class type checking to be more robust.  Other items will require more dicussion.  For example, I don't agree with the perspective of eliminating data.frame for the set_coverage function -- for taxonomy, a lot of people will already have a table of the taxon info in csv or other formats, and there will often be tens or hundreds of rows, so I think the function should take a table as input.  We could make it more flexible by allowing people to indicate which columns contain which ranks (as we do in the `codyn` package for a similar issue), but it still should be in tabular format.  More items like that to discuss, but they can be handled as improvements.  It will be much more accessible on CRAN.
