### 2016-11-11 ropensci/EML#183 @masalmon
Really great that the package is getting reviewed!

FWIW I thought it was actually handy to create a csv for the attributes. I have a set of tables so my csv also has a column "table".

I have two other points I've been thinking about recently, sorry if this is the wrong thread to mention them:

* Something I've been thinking of was the visualization of eml. If I remember correctly, when using the "official" software (KNB, dataONE?) one can navigate the xml? I wonder whether it'd be worth/possible to allow that as a RStudio addin.

* I'm going to write code for internal use at work for being able to use the eml for "translating" coded factor variables into their meaning (e.g. when in the data 1 = often and 2 = sometimes, we have that a lot in our dataset). Not sure whether this would be useful for other people?
