### 2016-11-11 #80 @gmbecker
# Further reviewer comment

@cboettig has started a discussion regarding the time-scale of implementing improvements vs putting out a stable release here ropensci/EML#183. In my capacity as a reviewer I feel it is most appropriate that I place my thoughts on this subject here as a secondary part of my review. Package authors, please find my feelings on this question below.

The most important aspect of a stable (read: CRAN) release is that the API is stable upon release for at least the near and preferably the medium term. As such, the most important issues to fix before release are ones that would actually break API compatibility. These are

- changing the output class of eml_get, and
- changing the names of existing constructors to remove the set_ prefix which I argue is misleading/a misnomer

This is because if you release  now without those changes, then release the next release with the changes in a half a year (or whenever), *people's existing code will break.* This is less acceptable for a CRAN release than it is for a package under heavy development within the ROpenSci incubator.

Beyond that, I feel the class checking is the next most important. People in the CRAN sphere are almost sure (in my opinion) to want to pass tbl_df objects to things that take data.frames. Tests to ensure that this works would not be out-of-place, and they should pass before CRAN release.

The other changes, while more important from a package design improvement perspective, can probably wait if they must, because they will likely be either backwards compatible or large enough that they warrant a new major version of the software.
