### 2016-11-11 ropensci/EML#183 @mbjones
@masalmon What do you mean by 'visualize'? The function lets you display the EML in XML format, and the review suggested that we should use something more R-like to show it hierarchically -- and keep our separate function for listing the XML version.  On the KNB and DataONE web sites, we translate the EML to an HTML version that presents it in a more readable format.  Is that what you mean?  Are you envisioning something similar within the R terminal?  Or launching a browser from the R terminal to view local metadata?  Just trying to understand your goal here.
