### 2016-11-11 ropensci/EML#183 @amoeba
I was thinking @masalmon was thinking of something like [listviewer](https://github.com/timelyportfolio/listviewer):

![viewer](https://cloud.githubusercontent.com/assets/563/20230527/faf734b8-a808-11e6-92bc-5e976dd7fc7e.gif)

It could be done as an RStudio Addin but also an optional part of the EML package via a function like `eml_view()`, and when called it'd show up in the RStudio Viewer pane as an interactive widget.
