### 2016-11-12 ropensci/EML#183 @masalmon
Yep @amoeba is right 😊
