+++ man/eml_locate_schema.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_validate.R
+\name{eml_locate_schema}
+\alias{eml_locate_schema}
+\title{eml_locate_schema}
+\usage{
+eml_locate_schema(eml)
+}
+\arguments{
+\item{eml}{an XML::XMLInternalDocument instance for an EML document}
+}
+\value{
+fully qualified path to the XSD schema for the appropriate version of EML
+}
+\description{
+eml_locate_schema returns the location of the XSD schema file for a given
+EML document, as shipped with the EML R package.
+}
+\details{
+Schema documents are copies of the schemas from the EML versioned
+releases. If an appropriate schema is not found, the function stops.
+}
+\examples{
+\donttest{
+schema <- eml_locate_schema(eml)
+}
+}
+
