+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(eml_find)
 export(eml_get)
+export(eml_locate_schema)
 export(eml_validate)
 export(get_TextType)
 export(get_attributes)
@@ -19,6 +20,8 @@ export(set_unitList)
 export(write_eml)
 import(XML)
 import(methods)
+importFrom(stringr,str_c)
+importFrom(stringr,str_match)
 importFrom(tools,file_ext)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
