### 2016-12-17 687f389
Added stringr dependency,and updated name.
