### 2016-12-17 ac157e4
Improved error handling in the eml_validate function.
Using try and tryCatch allows the function to return FALSE when an error would normally cause the function to stop.  Errors include being unable to find a suitable schema document, or having an invalid schema document. Related to issue #185.
