### 2016-12-19 61e0b6a
add cran comments for release
