+++ vignettes/metadata-repositories.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Metadata Repositories"
+title: "Metadata-Based Repositories"
 author: "Carl Boettiger"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteIndexEntry{Metadata-based repositories}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
