### 2016-12-20 099bdab
include gh-pages version of dcos from pkgdown
- Also modifies Description with URL for EML
- and updates Readme not to say no CRAN version is available.
