+++ DESCRIPTION
@@ -12,8 +12,9 @@ Authors@R: c(
 URL: https://github.com/ropensci/EML
 BugReports: https://github.com/ropensci/EML/issues
 Maintainer: Carl Boettiger <cboettig@gmail.com>
-Description: Parse and serialize Ecological Metadata Language ('EML') files into
-    S4 objects.
+Description: Parse and serialize Ecological Metadata Language ('EML', see
+  <https://en.wikipedia.org/wiki/Ecological_Metadata_Language> for
+  more information) files into S4 objects.
 License: FreeBSD
 LazyData: TRUE
 RoxygenNote: 5.0.1
