### 2016-12-20 ropensci/EML#181 @cboettig
v1.0.1 on CRAN now. http://cran.r-project.org/web/packages/EML
