### 2016-12-20 edf68d0
Add additional badges post-release
- CRAN badge
- CRAN downloads
- Zenodo DOI

Update docs, bump to dev version 1.0.1.1
