### 2016-12-20 #80 @cboettig
Update:

An initial version of the package is now on CRAN, after discussions with the `EML` package development team. The package is already in significant use by the DataONE folks, so it made sense to have a stable, archived version on CRAN. We hope to address the remainder of the issues highlighted here for a v1.1 release (https://github.com/ropensci/EML/milestone/9), which should address the remaining issues raised here that involve the higher-level interface, as well as a few performance issues reported by users.
