### 2016-12-20 ropensci/EML#181 @masalmon
Well done!
