### 2016-12-20 ropensci/EML#181 @gothub
Nicely done!

On Tue, Dec 20, 2016 at 11:41 AM, Maëlle Salmon <notifications@github.com>
wrote:

> Well done!
>
