### 2016-12-20 ropensci/EML#181 @mbjones
Awesome, Carl.  Glad it went smoothly!
