### 2017-01-19 6bd927d
adds tests for eml_view
