### 2017-01-19 fcf17a2
adds eml_view in the vignette about parsing EML
