+++ vignettes/parsing-EML.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Parsing EML"
+title: "Reading and parsing EML"
 author: "Carl Boettiger"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
@@ -9,6 +9,19 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+
+# Crude view of an EML file
+
+The function `eml_view()` allows, when the `listviewer` package is installed, to get a crude view of an EML file in the viewer. It can be useful for exploring the file.
+
+```{r, eval = FALSE}
+library("EML")
+f <- system.file("xsd/test/eml-i18n.xml", package = "EML")
+eml <- eml_view(f)
+
+```
+
+# Parsing an EML file
 
 The function `eml_get()` extracts a list of _all occurances_ the desired metadata elements from a given `eml` object or part thereof.  
 
@@ -37,8 +50,6 @@ Any EML element can be extracted in this way. Let's try an example metadata file
 eml_HARV <- read_eml("http://harvardforest.fas.harvard.edu/data/eml/hf001.xml")
 ```
 
-
-
 How many `dataTable` entities are there in this dataset?
 
 ```{r}
@@ -62,7 +73,7 @@ urls <- sapply(dt, function(x) x@physical[[1]]@distribution[[1]]@online@url)
 
 this verbose syntax can be useful if there are multiple `url` elements in each `dataTable` metadata, and we are trying to get only certain ones and not others. Specifying the exact path in this way can also improve the speed of the command.  For these reasons, programmatic use should consider this format, while the much simpler `eml_get` example shown above is practical for most interactive applications.
 
-All though the default return type for `eml_get` is just the S4 object (whose `print` method displays the corresponding XML structure used to represent that metadata), for a few commonly accessed complex elements, `eml_get` returns a more convenient `data.frame`.  For instance, the `attributeList` describing the metadata for every column in an EML document is returned as a pair of `data.frame`s, one for all the attributes, and an second optional `data.frame` defnining the levels for the factors, if any are used.  Let's take a look: 
+Although the default return type for `eml_get` is just the S4 object (whose `print` method displays the corresponding XML structure used to represent that metadata), for a few commonly accessed complex elements, `eml_get` returns a more convenient `data.frame`.  For instance, the `attributeList` describing the metadata for every column in an EML document is returned as a pair of `data.frame`s, one for all the attributes, and an second optional `data.frame` defnining the levels for the factors, if any are used.  Let's take a look: 
 
 Here we get the `attributeList` for each `dataTable` in the dataset. We check the length to confirm we get one `attributeList` for each `dataTable`
 
