### 2017-02-08 d71849a
Implement get_abstract and abstract part of get_TextType
As discussed in #194, we decided to implement a helper function called
get_abstract that would return a text representation of the EML
abstract. My first attempt involved copying most of the function body
of get_TextType into get_abstract so I refactored the two functions to
make use of a node_to_file helper method that does the conversion with
pandoc.
