+++ R/get_TextType.R
@@ -15,15 +15,15 @@
 #'
 #' @examples
 #' \donttest{
-#' 
+#'
 #' ## Convert an EML abstract to markdown
 #' f <- system.file("examples/hf205.xml", package = "EML")
 #' eml <- read_eml(f)
-#' abstract <- eml_get(eml, "abstract") 
+#' abstract <- eml_get(eml, "abstract")
 #' get_TextType(abstract[[1]], "markdown", "abstract.markdown")
 #' readLines("abstract.markdown")
 #' unlink("abstract.markdown") # tidy up
-#' 
+#'
 #' ## Turn a docx file into EML abstract and preview at HTML
 #' f <- system.file("examples/hf205-abstract.docx", package = "EML")
 #' a <- as(set_TextType(f), "abstract")
@@ -35,13 +35,31 @@ get_TextType <-
            to = "html",
            output = tempfile(class(node), fileext = paste0(".", to)),
            view = TRUE) {
-    
+
+    node_to_file(node, to, output)
+
+    if (view && to == "html")
+      utils::browseURL(basename(output))
+  }
+
+#' Converts a node into a file by running it through Pandoc
+#'
+#' @param node any node that supports a docbook-like structure (abstract, methods, etc)
+#' @param to desired format, default is html, but can be any type supported by pandoc (docx, md, etc)
+#' @param output name of the desired output file
+#'
+#' @return creates a file requested.
+#' @export
+#'
+#' @examples
+node_to_file <-
+  function(node, to = "html", output = tempfile(class(node), fileext = paste0(".", to))) {
     ## eml_get returns a list of nodes always.
-    if(!isS4(node) && length(node) == 1) 
+    if(!isS4(node) && length(node) == 1)
       node <- node[[1]]
-    
+
     # serialize sections in ListOfsection or paras from ListOfpara into XML document, save, rmarkdown into desired format
-    x <- XML::xmlChildren(S4Toeml(node))
+    x <- XML::xmlChildren(EML:::S4Toeml(node))
 
     if (!requireNamespace("rmarkdown", quietly = TRUE)) {
       stop("rmarkdown package required to convert to Docbook format",
@@ -50,34 +68,82 @@ get_TextType <-
     pandoc_convert <-
       getExportedValue("rmarkdown", "pandoc_convert")
 
-    wd <- getwd()
-    dir <- tempdir()
-    file.copy(output, file.path(dir, basename(output)), overwrite = TRUE)
-    setwd(dir)
-    docbook_file <- tempfile(pattern = "docbook", tmpdir = ".", fileext = ".db")
-    
+    docbook_path <- tempfile(pattern = "docbook", fileext = ".db")
     doctype <- XML::Doctype(name = "section",
-                       public = c("-//OASIS//DTD DocBook XML V4.2//EN"),
-                       system = "http://oasis-open.org/docbook/xml/4.5/docbookx.dtd")
+                            public = c("-//OASIS//DTD DocBook XML V4.2//EN"),
+                            system = "http://oasis-open.org/docbook/xml/4.5/docbookx.dtd")
     y <- XML::newXMLNode("article", x)
-    XML::saveXML(XML::xmlDoc(y), 
-                 docbook_file, doctype = doctype)
+    XML::saveXML(XML::xmlDoc(y),
+                 docbook_path,
+                 doctype = doctype)
+
     pandoc_convert(
-      basename(docbook_file),
-      to = to,
+      docbook_path,
       from = "docbook",
+      to = to,
       output = normalizePath(output, winslash = "/", mustWork = FALSE),
       options = "-s"
     )
 
-    file.copy(output, file.path(wd, basename(output)), overwrite = TRUE)
+    output
+  }
 
+#' Get the text content of the abstract
+#'
[…]
