### 2017-02-08 3dafa16
Fix up docs in get_TextType.R and related functions
