### 2017-03-02 39cbdb6
example of what R6-based EML classes might look like
