### 2017-03-03 57873c1
method to write out via xml2
