+++ R/classes.R
@@ -13,7 +13,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'value'))) {
       'not all elements are value objects'
-      
+
     } else {
       TRUE
     }
@@ -35,7 +35,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'temporalCoverage'))) {
       'not all elements are temporalCoverage objects'
-      
+
     } else {
       TRUE
     }
@@ -47,7 +47,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'geographicCoverage'))) {
       'not all elements are geographicCoverage objects'
-      
+
     } else {
       TRUE
     }
@@ -59,7 +59,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'taxonomicCoverage'))) {
       'not all elements are taxonomicCoverage objects'
-      
+
     } else {
       TRUE
     }
@@ -123,7 +123,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'section'))) {
       'not all elements are section objects'
-      
+
     } else {
       TRUE
     }
@@ -135,7 +135,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'para'))) {
       'not all elements are para objects'
-      
+
     } else {
       TRUE
     }
@@ -147,7 +147,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'itemizedlist'))) {
       'not all elements are itemizedlist objects'
-      
+
     } else {
       TRUE
     }
@@ -159,7 +159,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'orderedlist'))) {
       'not all elements are orderedlist objects'
-      
+
     } else {
       TRUE
     }
@@ -171,7 +171,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'emphasis'))) {
       'not all elements are emphasis objects'
-      
+
     } else {
       TRUE
     }
@@ -183,7 +183,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'subscript'))) {
       'not all elements are subscript objects'
-      
+
     } else {
       TRUE
     }
@@ -195,7 +195,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'superscript'))) {
       'not all elements are superscript objects'
-      
+
     } else {
       TRUE
     }
@@ -207,7 +207,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'literalLayout'))) {
       'not all elements are literalLayout objects'
-      
+
     } else {
       TRUE
     }
@@ -219,7 +219,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'citetitle'))) {
       'not all elements are citetitle objects'
-      
+
     } else {
       TRUE
     }
@@ -231,7 +231,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'ulink'))) {
       'not all elements are ulink objects'
-      
+
     } else {
       TRUE
     }
@@ -243,7 +243,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'listitem'))) {
       'not all elements are listitem objects'
-      
+
     } else {
       TRUE
     }
@@ -349,7 +349,7 @@ setClass(
   validity = function(object) {
     if (!all(sapply(object, is, 'individualName'))) {
       'not all elements are individualName objects'
-      
+
     } else {
       TRUE
     }
@@ -361,7 +361,7 @@ setClass(
[…]
