+++ man/write_eml.Rd
@@ -4,12 +4,12 @@
 \alias{write_eml}
 \title{write_eml}
 \usage{
-write_eml(eml, file = NULL, namespaces = NULL, ns = "eml", ...)
+write_eml(eml, file, namespaces = NULL, ns = "eml", ...)
 }
 \arguments{
 \item{eml}{an eml class object}
 
-\item{file}{file name to write XML. If NULL, will return XML as character string.}
+\item{file}{file name to write XML.}
 
 \item{namespaces}{named character vector of additional XML namespaces to use.}
 
@@ -29,4 +29,3 @@ f <- system.file("xsd/test", "eml.xml", package = "EML")
 eml <- read_eml(f)
 write_eml(eml)
 }
-
