+++ man/set_physical.Rd
@@ -59,4 +59,3 @@ set_physical
 \examples{
 set_physical("hf205-01-TPexp1.csv")
 }
-
