### 2017-03-03 a040b9b
textType test, format
