### 2017-03-04 2148bb4
WIP on migration to xml2 trial implementation for xml to s4
