+++ R/read_eml.R
@@ -4,16 +4,15 @@
 #'
 #' read_eml
 #' @param file path to an EML file
-#' @param ... additional arguments to \code{\link{xmlParse}}
+#' @param ... additional arguments to \code{\link{read_xml}}
 #' @return an eml object (S4 object)
 #' @export
-# @importFrom XML xmlRoot xmlParse
-#' @import methods XML
+#' @import methods xml2
 #' @examples
 #' f <- system.file("xsd/test", "eml.xml", package = "EML")
 #' eml <- read_eml(f)
 read_eml <- function(file, ...) {
-  node <- XML::xmlRoot(XML::xmlParse(file, ...))
+  node <- xml2::read_xml(file, ...)
   emlToS4(node)
 }
 
@@ -21,7 +20,7 @@ read_eml <- function(file, ...) {
 
 ## Creates a 'show' method so that eml S4 elements display in XML format instead of the
 ## impossible-to-read S4 format
-setMethod("show", signature("eml-2.1.1"), function(object)
-  show(S4Toeml(object)))
+#setMethod("show", signature("eml-2.1.1"), function(object)
+#  show(S4Toeml(object)))
 
 # cat(yaml::as.yaml(XML::xmlToList(S4Toeml(object))))
