+++ man/write_eml.Rd
@@ -25,7 +25,9 @@ If file is not specified, the result is a character string containing
 write_eml
 }
 \examples{
-f <- system.file("xsd/test", "eml.xml", package = "EML")
+f <- system.file("examples", "example-eml-valid.xml", package = "EML")
 eml <- read_eml(f)
-write_eml(eml)
+write_eml(eml, "test.xml")
+unlink("test.xml") # clean up
 }
+
