+++ man/set_coverage.Rd
@@ -64,3 +64,4 @@ coverage <-
               north = 37.38, south = 30.00)
 
 }
+
