### 2017-03-06 17b1db1
debugging xml2 write method
