+++ R/s4_to_xml.R
@@ -10,40 +10,136 @@ excluded_slots <- c("namespaces", "dirname", "slot_order", "xmlNodeName")
 #' @import methods
 s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
 
-    node_name <- class(obj)[[1]]
-    fields <- setdiff(slotNames(obj), excluded_slots)
+  node_name <- class(obj)[[1]]
+  fields <- setdiff(slotNames(obj), excluded_slots)
 
-    if(is.null(root)){
-      root <- xml_new_root(node_name,
-                           "xmlns:eml" = "eml://ecoinformatics.org/eml-2.1.1",
-                           "xmlns:xs" = "http://www.w3.org/2001/XMLSchema",
-                           "xmlns:xsi" = "http://www.w3.org/2001/XMLSchema-instance",
-                           "xmlns:stmml" = "http://www.xml-cml.org/schema/stmml_1.1")
-      xml <- root
-    } else {
-      xml <- xml_add_child(root, node_name)
-    }
+  who <- slotNames(obj)
+  if ("namespaces" %in% who & length(ns) == 0) ns <- obj@namespaces
+  slot_classes <- get_slots(class(obj))
+  is_attribute <- sapply(slot_classes, function(x) x == "xml_attribute")
+  attribute_elements <- who[is_attribute]
 
-    lapply(fields, function(child){
-      node <- slot(obj, child)
-      if(is.null(node) || emptynode(node)){
-        xml
-      } else if(is(node, "xml_attribute")){             # node is an attribute
-        if(child == "schemaLocation") child <- paste0("xsi:", child)  #Hack, should fix slot name to keep prefix
-        xml_set_attr(xml, child, as.character(node))
-      } else if(grepl("ListOf", class(node))){
-        lapply(node, s4_to_xml, xml)
-      } else if(is.basic(node)){
-        xml_set_text(xml, node)
-      } else if(isS4(node)){                            # node is S4 class
-        s4_to_xml(node, xml)
-      } else {                                          # node is a simple type
-        xml_add_child(xml, child, node)
+  ## clear namespace. except for stmml
+  if (!grepl("^stmml:", node_name)) node_name <- gsub("^[a-z]*:", "", node_name)
+  #if (length(ns) > 0) node_name <- paste0("eml:", node_name)
+
+
+  ## Okay, here we go, time to create some xml. First we'll need a root if we don't have it:
+  if(is.null(root)){
+    root <- xml_new_root(node_name,
+                         "xmlns:eml" = "eml://ecoinformatics.org/eml-2.1.1",
+                         "xmlns:xs" = "http://www.w3.org/2001/XMLSchema",
+                         "xmlns:xsi" = "http://www.w3.org/2001/XMLSchema-instance",
+                         "xmlns:stmml" = "http://www.xml-cml.org/schema/stmml_1.1")
+    xml <- root
+  } else {
+    xml <- xml_add_child(root, node_name)
+  }
+
+  base_attributes <- c("lang")
+  schema_attributes <- c("schemaLocation")
+
+
+  ## Handle the special case of "InlineType" -- which contain raw xml (for docbook text)
+  if (is(obj, "InlineType")) {  # should be inherits(obj, "InlineType") ?
+    xml_add_child(root, obj)  ## really? maybe need to subset the thing first?
+  } else {
+
+    who <- who[!(who %in% excluded_slots)] # drop excluded slots
+    names(who) <- who
+    ## Re-order 'who' to obey 'slot_order' slot.
+    ordering <- slot(obj, "slot_order")
+    who <- c(ordering, who[!who %in% ordering])
+
+    for (s in who) {
+      ## Attributes
+      if (s %in% attribute_elements) {
[…]
