+++ R/xml-s4.R
@@ -1,269 +1,0 @@
-## Used as the basis for all coercion methods from XML::XMLInternalNode
-## to EML's S4 objects.
-
-## Assumes slot is named by the name of the repeated element
-## but has type "ListOf<element>" which contains a list containing
-## the repeated elements.
-
-#' @import XML
-
-
-emlToS4 <- function (node, obj = new(xmlName(node)), ...) {
-  node_name <- fix_protected_names(xmlName(node))
-  attrs <- xmlAttrs(node)
-  children <- drop_comment_nodes(xmlChildren(node))
-  xml_names <- names(children)
-  xml_attr_names <- names(attrs)
-
-  s4 <- new(node_name)
-  slot_classes <- get_slots(node_name)
-  s4_names <- names(slot_classes)
-  subclasses <- xml_names[!xml_names %in% s4_names]
-  not_sub <- xml_names[xml_names %in% s4_names]
-  subattrs <- xml_attr_names[!xml_attr_names %in% s4_names]
-  not_subattrs <- xml_attr_names[xml_attr_names %in% s4_names]
-  subclasses <- c(subclasses, subattrs)
-  metanames <- s4_names[grepl("^[A-Z]", s4_names)]
-  metaclasses <- lapply(metanames, get_slots)
-  names(metaclasses) <- metanames
-
-  text_classes <- "text" %in% subclasses
-  subclasses <- subclasses[subclasses != "text"]
-
-  for (child in not_subattrs) {
-    slot(s4, child) <- new("xml_attribute", attrs[[child]])
-  }
-
-  if (inherits(s4, "InlineType")) {
-    kids <- xmlChildren(node)
-    class(kids) <- c("list", "XMLInternalNodeList")
-    s4@.Data  <- kids
-  } else {
-    if (length(metaclasses) == 0 && text_classes) {
-      #if("value" %in% names(children) && "value" %in% s4_names){
-      # s4@value <- listof(children, "value")
-      s4@.Data <-
-        paste(sapply(children[names(children) == "text"], xmlValue), collapse = "\n")
-      #} else {
-      #s4@.Data <- xmlValue(node)
-      #}
-    }
-
-
-    for (child in unique(subclasses)) {
-      y <- lapply(metaclasses, function(x)
-        match(child, names(x)))
-      s <- names(y)[!is.na(y)]
-      cls <- metaclasses[[s]][[y[[s]]]]
-      if (is.null(slot(s4, s)))
-        slot(s4, s) <- new(s)
-      if (grepl("^ListOf", cls))
-        slot(slot(s4, s), child) <- listof(children, child)
-      else if (cls == "character")
-        slot(slot(s4, s), child) <- xmlValue(children[[child]])
-      else if (cls == "xml_attribute")
-        slot(slot(s4, s), child) <-
-        new("xml_attribute" , attrs[[child]])
-      else
-        slot(slot(s4, s), child) <- emlToS4(children[[child]])
-    }
-
-    ## These are the normal s4@slot items
-    for (child in unique(not_sub)) {
-      cls <- slot_classes[[child]]
-      if (grepl("^ListOf", cls)) {
-        slot(s4, child) <- listof(children, child)
-      } else if (cls == "character") {
-        slot(s4, child) <- xmlValue(children[[child]])
-      } else {
-        slot(s4, child) <- emlToS4(children[[child]])
-      }
-    }
-
-
-  }
-
-  s4
-}
-
-
-#' @import XML
-#' @import methods
-S4Toeml <- s4_to_xml
-
-
-  oldfn <- function(obj,
-                    node = NULL,
-                    excluded_slots = c("namespaces", "dirname", "slot_order", "xmlNodeName"),
[…]
