### 2017-03-07 1e37bb2
parse and serialize & validate some simple test cases when using xml2
