+++ R/s4_to_xml.R
@@ -5,6 +5,8 @@ emptynode <- function(node){
 
 is.basic <- function(x) !isS4(x) && length(x) > 0 && (is.character(x) || is.numeric(x))
 excluded_slots <- c("namespaces", "dirname", "slot_order", "xmlNodeName")
+base_attributes <- c("lang")
+schema_attributes <- c("schemaLocation")
 
 #' @import xml2
 #' @import methods
@@ -36,13 +38,11 @@ s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
     xml <- xml_add_child(root, node_name)
   }
 
-  base_attributes <- c("lang")
-  schema_attributes <- c("schemaLocation")
 
 
   ## Handle the special case of "InlineType" -- which contain raw xml (for docbook text)
   if (is(obj, "InlineType")) {  # should be inherits(obj, "InlineType") ?
-    xml_add_child(root, obj)  ## really? maybe need to subset the thing first?
+    xml_add_child(xml, obj)  ## really? maybe need to subset the thing first?
   } else {
 
     who <- who[!(who %in% excluded_slots)] # drop excluded slots
@@ -61,35 +61,35 @@ s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
           if (s %in% schema_attributes)
             s <- paste0("xsi:", s)
           names(attrs) <- s
-          xml_set_attrs(root, attrs)
+          xml_set_attrs(xml, attrs)
         }
         ## metaTypes are used in slots rather than interitence to preserve element ordering!
         ## Capitalized slots are meta-types, and should not create a new xmlNode but instead
         ## pass their children directly to their parent node.
       } else if (grepl("^[A-Z]", s)) {
         X <- slot(obj, s)
-        xml_add_child(root, S4Toeml(X, root = root))
+        xml_add_child(xml, s4_to_xml(X, xml))
       } else {
         ## Complex child nodes
         X <- slot(obj, s)
           if (is(X, "InlineType"))
-            xml_add_child(root, xml_add_child(root, s, X))
+            xml_add_child(xml, xml_add_child(xml, s, X))
           else if (is(X, "list") && length(X)>0) {
             if (is(X[[1]], "InlineType"))
-              lapply(X, function(x)  xml_add_child(root, s, x))
+              lapply(X, function(x)  xml_add_child(xml, s, x))
             else if (is.character(X[[1]]) && length(get_slots(class(X[[1]]))) <= 1)
-              lapply(X, function(x) xml_add_child(root, class(x), x))
+              lapply(X, function(x) xml_add_child(xml, class(x), x))
             else
-              lapply(X, S4Toeml, root)
+              lapply(X, s4_to_xml, xml)
           } else if (isS4(X)) {
-            xml_add_child(root, S4Toeml(slot(obj, s)))
+            xml_add_child(xml, s4_to_xml(slot(obj, s)))
             ## Simple Child nodes
           } else if (length(X) > 0) {
             if (s == class(obj)[1] || s == ".Data")
               # special case
-              xml_add_children(root, X)   #
+              xml_add_children(xml, X)   #
             else
-              xml_add_children(root, xml_add_child(root, s, X))
+              xml_add_children(xml, xml_add_child(xml, s, X))
             ## No child node
           }
 
@@ -101,11 +101,10 @@ s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
 
   root
 }
-S4Toeml <- s4_to_xml
 
 
 
-dummy <- function(){
+s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
   node_name <- class(obj)[[1]]
   fields <- setdiff(slotNames(obj), excluded_slots)
 
[…]
