### 2017-03-07 4ea4655
fix validation / tests
schema location should not assume "eml" is first namespace given.  tests should check attributes on returned object rather than messages; this avoids side-effects and is easier to compute on.
