+++ tests/testthat/test-validation.R
@@ -7,6 +7,7 @@ testthat::test_that("We return TRUE when validating valid documents", {
 
   testthat::expect_true(eml_validate(f))
   testthat::expect_message(eml_validate(f), NA)
+  testthat::expect_length(attr(eml_validate(f), "errors"), 0)
 
 
   f2 <-
@@ -14,20 +15,21 @@ testthat::test_that("We return TRUE when validating valid documents", {
                 "example-eml-valid-special-characters.xml",
                 package = "EML")
 
-  testthat::expect_true(eml_validate(f2, encoding = "latin1"))
-  testthat::expect_message(eml_validate(f2, encoding = "latin1"), NA)
+  v <- eml_validate(f2, encoding = "latin1")
+  testthat::expect_true(v)
+  testthat::expect_message(v, NA)
+  testthat::expect_length(attr(v, "errors"), 0)
+
 
   # Check that we properly validate each supported version of EML
   versions <- c("eml-2.1.1", "eml-2.1.0")
   for(version in versions) {
-      eml <- system.file("examples", paste0("example-", version, ".xml"), package = "EML")
-      testthat::expect_true(eml_validate(eml))
-      testthat::expect_message(eml_validate(eml), NA)
+      f <- system.file("examples", paste0("example-", version, ".xml"), package = "EML")
+      testthat::expect_true(eml_validate(f))
   }
 })
 
 testthat::test_that("We return TRUE when validating valid eml objects", {
-  library("XML")
 
   f <-
     system.file("examples", "example-eml-valid.xml", package = "EML")
@@ -50,6 +52,10 @@ testthat::test_that("We return FALSE and messages when validating invalid docume
   f <-
     system.file("examples", "example-eml-invalid.xml", package = "EML")
 
-  testthat::expect_false(eml_validate(f))
-  testthat::expect_message(eml_validate(f))
+  v <- eml_validate(f)
+  testthat::expect_false(v)
+  testthat::expect_length(attr(v, "errors"), 1)
+  #testthat::expect_message(eml_validate(f)) # errors now returned as attributes of the return object, which is easier to compute on
+
+
 })
