+++ tests/testthat/test-testfiles.R
@@ -1,6 +1,8 @@
 testthat::context("We can parse & serialize test files")
 
-library(XML)
+## FIXME rewrite to be independent of all this XML assumption crap
+
+library(xml2)
 
 
 xml_tests <- list.files("inst/xsd/test/", "eml-.*\\.xml")
@@ -9,52 +11,50 @@ xml_tests <-
   xml_tests[-which(xml_tests == "eml-unitDictionary.xml")]
 
 
-out <- lapply(xml_tests, function(xml) {
-  testthat::test_that(xml, {
-    f <- system.file(paste0("xsd/test/", xml), package = "EML")
-    node <- xmlRoot(xmlParse(f))
-    ## This can cause trouble if not namespaced, and is not required
-    XML::removeAttributes(node, .attrs = "xsi:schemaLocation")
-    ## Test: can we parse into S4?
-    s4 <- as(node, xmlName(node))
-    ## Test: can we transform back into XML?
-    element <- as(s4, "XMLInternalNode")
-    ## Test: is our XML still schema-valid?
-    ## preserve the namespace of the input
-    ns <- xmlNamespaces(node)
-    xmlNamespaces(element) <- ns
-    ids <- sapply(ns, `[[`, "id")
-    tmp <- which(ids  == "eml")
-    if (length(tmp) > 0) {
-      ns_1 <- ids[[tmp]]
-    } else {
-      ns_1 <- ns[[1]]$id
-    }
-    setXMLNamespace(element, ns_1)
-    saveXML(xmlDoc(element), "test.xml")
 
-    all_elements <- xpathSApply(xmlParse(f), "//*", xmlName)
-    all_elements2 <- xpathSApply(xmlDoc(element), "//*", xmlName)
+all_test_examples<- function(xml) {
+  testthat::test_that(xml, {
+
+    message(paste("testing", xml))
+
+    ## Read EML
+    f <- system.file(paste0("xsd/test/", xml), package = "EML")
+    eml <- read_eml(f)
+    testthat::expect_true(isS4(eml))
+
+    ## Because root element in many test files is not "eml" but some sub-class like "dataset" or "access", we need to be explicit about namespace
+    original <- xml2::read_xml(f)
+    namespaces <- xml_ns(original)
+    ns <- names(namespaces[1])
+
+    ## Write and Validate EML (handling explicit namespacing)
+    write_eml(eml, "unittest.xml", namespaces = namespaces, ns = ns)
+    v <- eml_validate("unittest.xml")
+    testthat::expect_true(v)
+
+
+    ## Check no elements were lost:
+    original <- xml_name( xml_find_all(read_xml(f), "//*") )
+    test <- xml_name(xml_find_all(xml2::read_xml("unittest.xml"), "//*") )
 
     ## identical number of elements
-    testthat::expect_identical(length(all_elements), length(all_elements2))
-    ## identical modulo order
-    testthat::expect_identical(sort(all_elements), sort(all_elements2))
+    testthat::expect_identical(length(original), length(test))
+    ## elements are identical, modulo ordering
+    testthat::expect_identical(sort(original), sort(test))
     ## strictly identical:
-    if (!(xml == 'eml-physical.xml'))
-      # Skip known error
-      testthat::expect_identical(all_elements, all_elements2)
-
-    ## Validate
-    v <- eml_validate("test.xml")
-    testthat::expect_equal(v$status, 0)
+#    if (!(xml == 'eml-physical.xml'))      # Skip known error
+      testthat::expect_identical(original, test)
 
     ## Clean up
-    unlink("test.xml")
+    unlink("unittest.xml")
+    })
+}
 
-  })
-})
[…]
