+++ R/zzz.R
@@ -1,5 +1,7 @@
 
 ## Run this to have list available without having to re-parse EML every time
 ## FIXME consider sysdata for this?
-standardUnits <- get_unitList()
+
+
+#standardUnits <- get_unitList()
 
