+++ R/get_unitList.R
@@ -23,7 +23,7 @@
 get_unitList <-
   function(x = read_eml(system.file("xsd/eml-2.1.1/eml-unitDictionary.xml", package = "EML"))) {
     unitList <- x
-    if (is(unitList, "XMLInternalNode")) {
+    if (is(unitList, "xml_node")) {
       unitList <- emlToS4(unitList)
     } else if (is(unitList, "metadata")) {
       unitList <- emlToS4(unitList[[1]])
