### 2017-03-09 ac497e3
fix pruning of empty nodes
