### 2017-03-10 7e3eee7
continue handling special cases to avoid parsing/serializing errors
