### 2017-03-10 657be3d
basic tests on node number and node order now passing for all test files.
