### 2017-03-10 4ba3965
tidy uo old XML types
