+++ man/get_TextType.Rd
@@ -40,4 +40,3 @@ get_TextType(a)
 }
 
 }
-
