+++ man/eml_validate.Rd
@@ -4,13 +4,15 @@
 \alias{eml_validate}
 \title{eml_validate}
 \usage{
-eml_validate(eml, encoding = "UTF-8")
+eml_validate(eml, encoding = "UTF-8", schema = NULL)
 }
 \arguments{
 \item{eml}{an eml class object, file, or xml document}
 
 \item{encoding}{optional, if eml is a file path / an eml and has special characters, one can
 gives the encoding used by xmlParse.}
+
+\item{schema}{path to schema}
 
 \item{...}{additional arguments to eml_write, such as namespaces}
 }
@@ -41,4 +43,3 @@ eml_validate(dataset, namespaces = c(ds = "eml://ecoinformatics.org/dataset-2.1.
 }
 
 }
-
