+++ man/set_attributes.Rd
@@ -47,4 +47,3 @@ For dateTime data:
 
 For factor data:
 }
-
