### 2017-03-14 c4f3afe
fixing validation issues
