### 2017-03-14 0709564
generalize validate to use ns matching schemaLocation
