### 2017-03-14 5f6b358
migrate set_TextType to xml2
