+++ R/set_TextType.R
@@ -53,19 +53,17 @@ set_TextType <- function(file = NULL, text = NULL) {
 
 
 set_section <- function(docbook) {
-  sections <-
-    XML::xpathApply(docbook, "/article/sect1", XML::xmlChildren)
-  s <- lapply(sections, function(x)
-    new("section", as(x, "list")))
+  sections <- lapply(xml2::xml_find_all(docbook, "/article/sect1"), xml2::xml_children)
+  s <- lapply(sections, function(x) new("section", list(x)))
   as(s, "ListOfsection")
 }
 
 
 
 set_para <-  function(docbook) {
-  para <- XML::xpathApply(docbook, "/article/para", XML::xmlChildren)
+  para <- xml2::xml_find_all(docbook, "/article/para")
   s <- lapply(para, function(x)
-    new("para", as(x, "list")))
+    new("para", list(x)))
   as(s, "ListOfpara")
 }
 
@@ -92,12 +90,12 @@ to_docbook <- function(file = NULL) {
       output = normalizePath(docbook_file, winslash = "/", mustWork = FALSE),
       options = "-s"
     )
-    docbook <- XML::xmlParse(docbook_file)
+    docbook <- xml2::read_xml(docbook_file)
     setwd(wd)
 
   } else {
     ## File is already xml/docbook, so no need for pandoc
-    docbook  <- XML::xmlParse(file)
+    docbook  <- xml2::read_xml(file)
   }
   docbook
 
