+++ NAMESPACE
@@ -26,6 +26,7 @@ importFrom(tools,file_ext)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
 importFrom(utils,person)
+importFrom(uuid,"UUIDgenerate()")
 importFrom(xml2,read_xml)
 importFrom(xml2,write_xml)
 importFrom(xml2,xml_ns)
