+++ DESCRIPTION
@@ -21,7 +21,8 @@ RoxygenNote: 6.0.1
 Imports:
     xml2 (>= 1.1.1),
     methods,
-    tools
+    tools,
+    uuid
 Suggests:
     testthat,
     knitr,
