+++ R/set_unitList.R
@@ -154,7 +154,7 @@ set_unitList <-
 
     if (as_metadata) {
       xml_meta <- S4Toeml(out)
-      set_xml_namespace(xml_meta,
+      xml_set_namespace(xml_meta,
                       c(stmml =  "http://www.xml-cml.org/schema/stmml_1.1"))
       new("additionalMetadata", metadata = new("metadata", list(xml_meta)))
     } else{
