+++ R/read_eml.R
@@ -20,7 +20,5 @@ read_eml <- function(file, ...) {
 
 ## Creates a 'show' method so that eml S4 elements display in XML format instead of the
 ## impossible-to-read S4 format
-#setMethod("show", signature("eml-2.1.1"), function(object)
-#  show(S4Toeml(object)))
-
-# cat(yaml::as.yaml(XML::xmlToList(S4Toeml(object))))
+setMethod("show", signature("eml-2.1.1"), function(object)
+  show(s4_to_xml(object)))
