+++ R/eml_view.R
@@ -1,10 +1,10 @@
 #' View an EML
 #'
 #' @param eml_path Path to the EML one wants to view
-#'
+#' @param ... additional arguments to \code{\link{read_xml}}
 #' @return Nothing, the EML is viewed in the Viewer.
 #' @export
-#'
+#' @importFrom xml2 as_list
 #' @details The function needs the \code{listviewer} package which you can install from CRAN
 #' via \code{install.packages("listviewer"))}. The function uses \code{listviewer::jsonedit}
 #' which gives a crude but useful viewing of the EML
@@ -20,17 +20,17 @@
 #'  eml_view(f)
 #'  f <- system.file("examples", "example-eml-2.1.0.xml", package = "EML")
 #'  eml_view(f)
-eml_view <- function(eml_path){
+eml_view <- function(eml_path, ...){
   if (!requireNamespace("listviewer", quietly = TRUE)) {
       message("eml_view needs the listviewer package, \n
               Install it via install.packages('listviewer')")
     return(NULL)
     }
 
-  eml_list <- XML::xmlTreeParse(eml_path, encoding = "latin1")
-  eml_list <- XML::xmlToList(eml_list,
-                             addAttributes = FALSE,
-                             simplify = FALSE)
+  doc <- read_xml(eml_path, ...)
+  eml_list <- xml2::as_list(doc)
 
-  listviewer::jsonedit(listdata = eml_list)
+  ## shouldn't call function from a Suggested package, see get_TextType for how we use rmarkdown fns
+  jsonedit <- getExportedValue("listviewer", "jsonedit")
+  # jsonedit(eml_list)
 }
