+++ R/get_unitList.R
@@ -14,7 +14,7 @@
 #' # Read in additional units defined in a EML file
 #' f <- system.file("xsd/test/eml-datasetWithUnits.xml", package = "EML")
 #' eml <- read_eml(f)
-#' unitList <- get_unitList(eml@@additionalMetadata[[1]]@@metadata[[1]])
+#' unitList <- get_unitList(eml@@additionalMetadata[[1]]@@metadata)
 #'
 #' ## Read in the definitions of standard units:
 #' get_unitList()
