+++ R/set_TextType.R
@@ -3,7 +3,6 @@
 #' For any EML element of class TextType, this function can be used to generate the appropriate EML from a markdown-formatted file.
 #' @param text a plain text character string which will be used directly as the content of the node if no file is given
 #' @param file path to a file providing formatted input text, see details.
-#' @import XML
 #' @return a TextType object that can be coerced into any element inheriting from TextType, see examples
 #' @importFrom tools file_ext
 #' @details If the `rmarkdown` package is installed, then the input file can
