+++ R/read_eml.R
@@ -20,5 +20,8 @@ read_eml <- function(file, ...) {
 
 ## Creates a 'show' method so that eml S4 elements display in XML format instead of the
 ## impossible-to-read S4 format
-setMethod("show", signature("eml-2.1.1"), function(object)
-  show(s4_to_xml(object)))
+setMethod("show", signature("eml-2.1.1"), function(object){
+  tmp <- tempfile()
+  write_eml(object, tmp)
+  cat(readLines(tmp), sep="\n")
+})
