### 2017-03-15 ecbd73a
Use formatted xml for default show method #182
