+++ R/read_eml.R
@@ -22,6 +22,6 @@ read_eml <- function(file, ...) {
 ## impossible-to-read S4 format
 setMethod("show", signature("eml-2.1.1"), function(object){
   tmp <- tempfile()
-  write_eml(object, tmp)
+  write_eml(object, tmp, namespaces = character(), ns = character(), options = c("format", "no_declaration"))
   cat(readLines(tmp), sep="\n")
 })
