### 2017-03-20 b09e49c
adds test for an EML that's just been created
