+++ tests/testthat/test-validation.R
@@ -69,3 +69,236 @@ testthat::test_that("We can handle validation against sub-schema alone (e.g. aga
 
 })
 
+
+testthat::test_that("We return TRUE when validating a valid just created eml object", {
+  attributes <-
+    data.frame(
+      attributeName = c(
+        "run.num",
+        "year",
+        "day",
+        "hour.min",
+        "i.flag",
+        "variable",
+        "value.i",
+        "length"),
+      attributeDefinition = c(
+        "which run number (=block). Range: 1 - 6. (integer)",
+        "year, 2012",
+        "Julian day. Range: 170 - 209.",
+        "hour and minute of observation. Range 1 - 2400 (integer)",
+        "is variable Real, Interpolated or Bad (character/factor)",
+        "what variable being measured in what treatment (character/factor).",
+        "value of measured variable for run.num on year/day/hour.min.",
+        "length of the species in meters (dummy example of numeric data)"),
+      formatString = c(
+        NA,
+        "YYYY",
+        "DDD",
+        "hhmm",
+        NA,
+        NA,
+        NA,
+        NA),
+      definition = c(
+        "which run number",
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA),
+      unit = c(
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        "meter"),
+      numberType = c(
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        NA,
+        "real"),
+      stringsAsFactors = FALSE
+    )
+
+  i.flag <- c(R = "real",
+              I = "interpolated",
+              B = "bad")
+  variable <- c(
+    control  = "no prey added",
+    low      = "0.125 mg prey added ml-1 d-1",
+    med.low  = "0,25 mg prey added ml-1 d-1",
+    med.high = "0.5 mg prey added ml-1 d-1",
+    high     = "1.0 mg prey added ml-1 d-1",
+    air.temp = "air temperature measured just above all plants (1 thermocouple)",
+    water.temp = "water temperature measured within each pitcher",
+    par       = "photosynthetic active radiation (PAR) measured just above all plants (1 sensor)"
+  )
+
+  value.i <- c(
+    control  = "% dissolved oxygen",
+    low      = "% dissolved oxygen",
+    med.low  = "% dissolved oxygen",
+    med.high = "% dissolved oxygen",
+    high     = "% dissolved oxygen",
+    air.temp = "degrees C",
+    water.temp = "degrees C",
+    par      = "micromoles m-1 s-1"
+  )
+
+  ## Write these into the data.frame format
+  factors <- rbind(
+    data.frame(
+      attributeName = "i.flag",
+      code = names(i.flag),
+      definition = unname(i.flag)
+    ),
+    data.frame(
+      attributeName = "variable",
+      code = names(variable),
+      definition = unname(variable)
+    ),
+    data.frame(
+      attributeName = "value.i",
+      code = names(value.i),
+      definition = unname(value.i)
+    )
+  )
+
+  attributeList <- set_attributes(attributes, factors, col_classes = c("character", "Date", "Date", "Date", "factor", "factor", "factor", "numeric"))
+
+
+  physical <- set_physical("hf205-01-TPexp1.csv")
+
+  dataTable <- new("dataTable",
+                   entityName = "hf205-01-TPexp1.csv",
+                   entityDescription = "tipping point experiment 1",
+                   physical = physical,
+                   attributeList = attributeList)
+
+  getSlots("dataTable")
+
+  example <- new("dataTable")
+  #example@physical <- physical # Error -- wrong type.
+  example@physical <- c(physical) # concat method promotes 1 or more arguments to a ListOf
+
+
[…]
