### 2017-03-20 de6aa4b
needs to save the file with encoding latin1 for the test to fail?
