### 2017-03-20 8a7ce5e
tries to fix the encoding issue on Travis
