+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(EML)
 
-test_check("EML")
+test_check("EML", encoding = "latin1")
