+++ appveyor.yml
@@ -14,6 +14,7 @@ install:
 
 build_script:
   - travis-tool.sh install_deps
+  - travis-tool.sh install_github hadley/testthat
 
 test_script:
   - travis-tool.sh run_tests
