### 2017-03-20 05b88a7
adds test for reproducing bug reported in #207
