### 2017-03-20 be0faa7
fix coercion for additional metadata
(needs to be a nodeset, not a node)
