### 2017-03-20 a9a0cd3
streamline test after merge
