### 2017-03-20 24fa6bd
Does appveyor like pandoc-tests now?
