### 2017-03-20 df741a3
fix external test file
