+++ inst/examples/create-full-eml.R
@@ -134,7 +134,7 @@ coverage <-
 methods_file <- system.file("examples/hf205-methods.docx", package = "EML")
 methods <- set_methods(methods_file)
 
-R_person <- as.person("Ma?l Ellison <fakeaddress@email.com> [cre]")
+R_person <- as.person("Maëlle Maëlle <fakeaddress@email.com> [cre]")
 aaron <-as(R_person, "creator")
 
 others <- c(as.person("Benjamin Baiser [ctb]"), as.person("Jennifer Sirota [ctb]"))
@@ -204,239 +204,6 @@ see: http://www.lternet.edu/data/netpolicy.html."
 
 abstract_file <-  system.file("examples/hf205-abstract.md", package = "EML")
 abstract <- as(set_TextType(abstract_file), "abstract")
-
-dataset <- new("dataset",
-               title = title,
-               creator = aaron,
-               pubDate = pubDate,
-               intellectualRights = intellectualRights,
-               abstract = abstract,
-               associatedParty = associatedParty,
-               keywordSet = keywordSet,
-               coverage = coverage,
-               contact = contact,
-               methods = methods,
-               dataTable = dataTable)
-
-eml <- new("eml",
-           packageId = "f0cda3bf-2619-425e-b8be-8deb6bc6094d",  # from uuid::UUIDgenerate(),
-           system = "uuid", # type of identifier
-           dataset = dataset)
-
-
-v <- eml_validate(eml, encoding = "latin1")
-testthat::expect_true(v)
-testthat::expect_message(v, NA)
-testthat::expect_length(attr(v, "errors"), 0)
-})
-
-testthat::test_that("We return TRUE when validating a valid just created eml object with unit list", {
-  attributes <-
-    data.frame(
-      attributeName = c(
-        "run.num",
-        "year",
-        "day",
-        "hour.min",
-        "i.flag",
-        "variable",
-        "value.i",
-        "length"),
-      attributeDefinition = c(
-        "which run number (=block). Range: 1 - 6. (integer)",
-        "year, 2012",
-        "Julian day. Range: 170 - 209.",
-        "hour and minute of observation. Range 1 - 2400 (integer)",
-        "is variable Real, Interpolated or Bad (character/factor)",
-        "what variable being measured in what treatment (character/factor).",
-        "value of measured variable for run.num on year/day/hour.min.",
-        "length of the species in terameters (dummy example of numeric data)"),
-      formatString = c(
-        NA,
-        "YYYY",
-        "DDD",
-        "hhmm",
-        NA,
-        NA,
-        NA,
-        NA),
-      definition = c(
-        "which run number",
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA),
-      unit = c(
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        "terameter"),
-      numberType = c(
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        NA,
-        "real"),
-      stringsAsFactors = FALSE
-    )
-
-  i.flag <- c(R = "real",
-              I = "interpolated",
-              B = "bad")
-  variable <- c(
-    control  = "no prey added",
-    low      = "0.125 mg prey added ml-1 d-1",
-    med.low  = "0,25 mg prey added ml-1 d-1",
-    med.high = "0.5 mg prey added ml-1 d-1",
-    high     = "1.0 mg prey added ml-1 d-1",
-    air.temp = "air temperature measured just above all plants (1 thermocouple)",
-    water.temp = "water temperature measured within each pitcher",
[…]
