### 2017-03-20 4c3390e
looks like #205 fixed pandoc issues of #179 as well :crossed_fingers:
