### 2017-03-21 0f7a2ac
experimenting with appveyor and travis settings
