+++ .travis.yml
@@ -8,9 +8,9 @@ r:
 warnings_are_errors: true
 sudo: required
 
-r_github_packages:
-- jimhester/covr
-- hadley/testthat
+#r_github_packages:
+#- jimhester/covr
+#- hadley/testthat
 
 after_success:
 - Rscript -e 'library(covr); codecov()'
