+++ appveyor.yml
@@ -10,11 +10,20 @@ init:
 install:
   ps: Bootstrap
 
+cache:
+  - C:\RLibrary
+
 # Adapt as necessary starting from here
+
+## Pandoc installation based on https://github.com/jennybc/reprex/blob/master/appveyor.yml
+before_test:
+  - cinst pandoc
+  - ps: $env:Path += ";C:\Program Files (x86)\Pandoc\"
+  - pandoc -v
 
 build_script:
   - travis-tool.sh install_deps
-  - travis-tool.sh install_github hadley/testthat
+#  - travis-tool.sh install_github hadley/testthat
 
 test_script:
   - travis-tool.sh run_tests
@@ -42,15 +51,3 @@ artifacts:
   - path: '\*_*.zip'
     name: Bits
 
-cache:
-  - C:\RLibrary
-  - C:\Program Files (x86)\Pandoc\
-
-# Adapt as necessary starting from here
-before_test:
-  - ps: >-
-      if (-Not (Test-Path "C:\Program Files (x86)\Pandoc\")) {
-        cinst pandoc
-      }
-  - ps: $env:Path += ";C:\Program Files (x86)\Pandoc\"
-  - pandoc -v
