### 2017-03-21 9d5af45
separate file for show methods
draft of what coffee-methods should look like
