### 2017-03-21 16b2315
separate out encoding tests
