### 2017-03-23 5b1a332
Merci @aammd for the useful unsollicited code review
