### 2017-03-24 8ccee9e
now runs the test on Windows
