### 2017-04-20 4dfbd7d
use canonical form link to CRAN in README
