+++ README.Rmd
@@ -18,7 +18,7 @@ knitr::opts_chunk$set(
 [![codecov.io](https://codecov.io/github/ropensci/EML/coverage.svg?branch=master)](https://codecov.io/github/ropensci/EML?branch=master)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/EML?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/EML)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/EML)](http://cran.r-project.org/package=EML)
-[![CRAN RStudio mirror downloads](http://cranlogs.r-pkg.org/badges/EML)](http://cran.r-project.org/web/packages/EML/index.html)
+[![CRAN RStudio mirror downloads](http://cranlogs.r-pkg.org/badges/EML)](https://CRAN.R-project.org/package=EML)
 [![Github Stars](https://img.shields.io/github/stars/ropensci/EML.svg?style=social&label=Github)](https://github.com/ropensci/EML)
 [![DOI](https://zenodo.org/badge/10894022.svg)](https://zenodo.org/badge/latestdoi/10894022)
 
