### 2017-04-25 859f566
canonical cran link should be https
