### 2017-04-25 b52acf6
streamline doc examples
