+++ man/eml_get.Rd
@@ -41,12 +41,10 @@ for example, compare \code{\link{get_attributes}} to the example shown here.
 }}
 
 \examples{
+\donttest{
 f <- system.file("xsd/test", "eml-datasetWithUnits.xml", package = "EML")
 eml <- read_eml(f)
 eml_get(eml, "physical")
-
-## slower examples
-\donttest{
 eml_get(eml, "attributeList")
 
 ## The first argument need not be an "eml" class, it could be a child element; e.g.
