### 2017-04-25 3a7c019
bump version, ready for release
