+++ NEWS.md
@@ -1,8 +1,9 @@
-# EML 1.0.2.1
+# EML 1.0.2
 
 * Migrate dependency from XML to xml2
-* Improved performance in displaying and serializing XML (#)[]
-* Added a `NEWS.md` file to track changes to the package.
+* Improved performance in displaying and serializing XML
+* Added a `NEWS.md` file to track changes to the package
+* Minor bug fixes
 
 
 # EML 1.0.1
