### 2017-05-01 9a0de77
use pandoc conditionally
Some CRAN platforms have rmarkdown but still do not have pandoc, so rmarkdown functions that really need pandoc fail.
