### 2017-05-01 4b04809
patch for CRAN machines lacking pandoc
