### 2017-05-04 b5ee51d
Add a check for combinations of measurementScale and domain
