### 2017-06-07 33d04e5
Fix a bug in serialization of species taxonomic rank
