### 2017-06-07 37686f6
Fix mistakes in set_taxonomicCoverage examples
