### 2017-06-07 8f9199c
Document the expand arg in set_taxonomicCoverage
