### 2017-06-07 d7ae40b
Minor doc tweaks on set_taxonomicCoverage
