### 2017-06-07 02a4c16
Update news.md
