### 2017-06-15 fdbeb82
Make set_taxonomicCoverage more robust
This request and code comes from John Porter jhp7e@eservices.virginia.edu. Thanks John!
