+++ R/set_coverage.R
@@ -185,11 +185,26 @@ set_taxonomicCoverage <- function(sci_names, expand=FALSE) {
 
     classifications <- taxize::classification(sci_names, db = 'itis')
 
-    sci_names <- lapply(classifications, function(cls) {
-      x <- as.list(cls[["name"]])
-      names(x) <- cls[["rank"]]
-      x
-    })
+    # Remove any NAs and warn for each
+    if (any(is.na(classifications))) {
+      warning(call. = FALSE,
+              paste0("Some scientific names were not found in the taxonomic database and were not expanded: ", paste0(sci_names[which(is.na(classifications))], collapse = ","), "."))
+    }
+
+    # Turn result into a list of named lists where names are the rank name and
+    # values are the rank value
+    sci_names <- mapply(function(cls, sci_name) {
+      # If the species name isn't found in the database, NA is returned
+      if (is.list(cls)) {
+        x <- as.list(cls[["name"]])
+        names(x) <- cls[["rank"]]
+        x
+      } else {
+        x <- list(list("species" = as.character(sci_name)))
+        names(x) <- sci_name
+        x
+      }
+    }, classifications, names(classifications), SIMPLIFY = FALSE)
   }
 
   if (class(sci_names) == "character") {
