### 2017-06-15 8217162
Fix a minor bug in set_taxonomicCoverage
