### 2017-06-15 6ee9f86
Add support in set_taxonomicCoverage for alternative dbs
This is (again) from John Porter (jhp7e@eservices.virginia.edu). Thanks John!
