### 2017-07-05 2bc6281
Add support in set_physical for automatic size/checksum calculation
Addresses #220
