### 2017-07-05 6f323f1
Add support in set_physical for automatic size/checksum calculation (#221)
* Add support in set_physical for automatic size/checksum calculation

Addresses #220

* Update Rd

* Add digest to Imports to fix R CMD CHECK
