### 2017-07-13 a06a439
Fix #189 issue (listviewer function commented out)
We'll see if this passes checks!
