### 2017-07-21 ropensci/EML#183 @cboettig
quickly addressable issues from the onboarding review:

- [ ] patch up the eml_get return types,
- [ ] fix some of the autogenerated code issues, and
- [ ] fix the class type checking to be more robust.
