+++ inst/onboarding/gmbecker-review.md
@@ -1,0 +1,120 @@
+## Package Review
+
+_Please check off boxes as applicable, and elaborate in comments below.  Your review is not limited to these topics, as described in the reviewer guide_
+- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).
+
+#### Documentation
+
+The package includes all the following forms of documentation:
+- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
+- [x] **Installation instructions:** for the development version of package and any non-standard dependencies in README
+- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally
+- [x] **Function Documentation:** for all exported functions in R help
+- [x] **Examples** for all exported functions in R Help that run successfully locally
+- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION 
+
+#### Functionality
+- [x] **Installation:** Installation succeeds as documented.
+- [x] **Functionality:** Any functional claims of the software been confirmed.
+- [x] **Performance:** Any performance claims of the software been confirmed.
+- [x] **Automated tests:** Unit tests cover essential functions of the package
+  and a reasonable range of inputs and conditions. All tests pass on the local machine.
+- [x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
+
+#### Final approval (post-review)
+- [ ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
+
+Estimated hours spent reviewing: 
+
+## 6 for initial review and comment writing
+
+# Review Comments
+
+I do not have experience within the field of application for this package. As such, I will focus in this review on the design and technical aspects of the package while speaking less to it's direct applicability to problems computational ecologists actually face. 
+
+The package seems generally good, with code (both generated and hand-written) which is well conceived and behaves as intended.  I feel that with afew revisions to tighten up some aspects of it, it will be a strong and welcome contribution to ROpenSci core mission to facilitate interaction with open-data sources and formats from within R.
+
+## S4 design and usage
+
+ there are some S4-specific design and stylistic choices which I recommend the authors revisit.
+
+### Constructors
+
+Generally, all classes which any end users are expected to create themselves should have exported constructors, rather than requiring users to call `new` directly. This preserves the abstraction if the underlying class structure changes, and usually increases the readability of the users' code.
+
+This is doable even when classes (and thus constructors) are automatically generated. I wrote some code that seems to successfully do the individual constructor generation for this package. I'm happy to contribute it, though achieving a positive final review and acceptance is, of course, in no way contingent on that particular code being used. 
+
+Another note about constructors is that I believe it is stylistically cleaner to map the arguments directly to the slots and to use constructors for the classes of those slots as necessary, recursively, rather than utilizing nested `new` calls as the code does now, e.g., in `set_physical`.
+
+Finally, I do not feel that set_ is an appropriate prefix for constructors, as in the S4 OOP style that suggests than an existing object is being modified by having a component (e.g., slot) of it set. The standard paradigm is to have constructors be functions with names identical to the classes they construct, although that is not mandatory.
+
+### show method
+
+The current show method for `eml` objects converts the object to XML before printing. This seems cumbersome, and, I suspect, is likely to cause problems for real-world-sized EML documents (in the same way the XML package does when printing a large XML blob).  It seems that EML is structured enough that a meaningful summary could be displayed by `show`, while  the emlToXML function would still allow users to easily display the entire XML blob if desired.  
+
+### duplicated classes
+
[…]
