+++ inst/onboarding/review-reply.md
@@ -1,0 +1,117 @@
+We believe we have now addressed the very helpful issues and suggestions raised by both reviewers.  While EML is still under active development that may address some of the more big-picture issues about being more user friendly, we feel this addresses the essential issues highlighted by reviewers as listed below.  Apologies for the long delay.  
+
+
+# cgries
+
+
+---
+
+
+> For the average user (me included) it would be nice to have a little more extensive documentation. I.e., in the description field of functions it would help to see more than just the function name again.  The examples are all very helpful and work great. The vignettes are great as well and really help to get started writing code to generate EML. The explanation for the ‘new’ method is, however, very short. I realize that it mainly takes understanding XML and EML specifically to get the hang of that and EML is complex with many ‘new’ constructs for using the more obscure elements.
+
+We agree.  We're working on making easier constructors, as suggested by @gmbecker, basically this is just waiting on his s4Coffee implementation which should be able to automate the generation of the 100s of methods needed to do this.  With help from the user community, we're also flushing out documentation aimed at users less familiar with EML and/or R internals, though of course it was necessary to get the fundamentals in place first, which is what we feel the current release now does.
+
+> One improvement that could still be considered for the first release is dealing with people information. I found myself very quickly writing some helper functions which I then also discovered in the arctic repository that was mentioned by @amoeba in [https://github.com/ropensci/EML/issues/183](url). Given the fact that we both seemed to need such and the basic `as.person` is not sufficient, some thought might go into improving that before everyone starts writing those helper functions. I would like to see the option to provide a csv file/data frame with the people information following the pattern of attribute and taxonomic coverage information. I have had datasets with almost 60 creators and their sequence changed several times in the process of publishing the paper. That would be very tedious to work out in code.
+Which leads me to responding to @gmbecker that I find it very convenient to provide attribute and taxonomic information as a data frame. It allows for easy retrieval of the information from either a database or more frequently from a spreadsheet that data providers supply with this information. Most data providers don’t know about the EML subtleties of measurementScale or how to correctly spell the units, but it takes very little editing on my part to prepare a spreadsheet from the information that I do receive. Taxonomic coverage can be extensive and again frequently comes out of a database, i.e., can be easily provided as csv file and data frame.
+
+We agree that being able to maintain metadata such as taxonomic information or author details in in simple .csv files that can quickly be read into EML is a big win, and have been building methods around these patterns.  Understanding the use cases such as @cgries outlines here helps us do that better.  This is all currently possible with the `new` constructors, but requires a bit more coding.  Once we've auto
+
+
+----
+
+# gmbecker
+
+
+## Package Review
+
+
+
+### Constructors
+
+> Generally, all classes which any end users are expected to create themselves should have exported constructors, rather than requiring users to call `new` directly. This preserves the abstraction if the underlying class structure changes, and usually increases the readability of the users' code. ....
+
+We agree, this will be done once we can automate the creation of these methods with your awesome S4Coffee package!  See https://github.com/gmbecker/S4Coffee/issues/1.  
+
+### show method
+
+> The current show method for `eml` objects converts the object to XML before printing. This seems cumbersome, and, I suspect, is likely to cause problems for real-world-sized EML documents (in the same way the XML package does when printing a large XML blob).  It seems that EML is structured enough that a meaningful summary could be displayed by `show`, while  the emlToXML function would still allow users to easily display the entire XML blob if desired.
+
+Yup, display issues now addressed in #182 and #189.  
+
+### duplicated classes
+
[…]

[…]
