### 2017-08-16 7c3c501
@importFrom xml2 instead of @import xml2
