+++ R/s4_to_xml.R
@@ -49,23 +49,23 @@ s4_to_xml <- function(obj, root = NULL, ns = eml_namespaces){
     } else if(is(node, "xml_attribute")){             # node is an attribute
       if(child == "schemaLocation") child <- paste0("xsi:", child)  #Hack, should fix slot name to keep prefix. (schema_attributes)
       if(child %in% base_attributes)  child <- paste0("xml:", child)
-      xml_set_attr(xml, child, as.character(node))
+      xml2::xml_set_attr(xml, child, as.character(node))
     } else if(grepl("^[A-Z][a-z]", child)){                # node is a metanode (class whose children should all become slots)
       s4_to_xml(node, xml)
     } else if (is(node, "InlineType")  && length(node) > 0 && is(node[[1]], "xml_nodeset")){
-      xml <- xml_add_child(xml, child)
-      lapply(node[[1]], function(n) xml_add_child(xml, n))
+      xml <- xml2::xml_add_child(xml, child)
+      lapply(node[[1]], function(n) xml2::xml_add_child(xml, n))
     ## FIXME is this next case no longer needed? is the same as the case above?
     } else if(is(node, "list") && length(node) > 0 && is(node@.Data[[1]], "xml_nodeset")){
       lapply(node@.Data[[1]], function(n) xml_add_child(xml, n))
     } else if(grepl("ListOf", class(node))){
       lapply(node, s4_to_xml, xml)
     } else if(is.basic(node) && child == ".Data"){
-      xml_set_text(xml, node)
+      xml2::xml_set_text(xml, node)
     } else if(isS4(node)){                            # node is S4 class
       s4_to_xml(node, xml)
     } else if(child != ".Data"){                                          # node is a simple type
-      xml_add_child(xml, child, node)
+      xml2::xml_add_child(xml, child, node)
     }
   })
 
