+++ NAMESPACE
@@ -20,7 +20,6 @@ export(set_taxonomicCoverage)
 export(set_unitList)
 export(write_eml)
 import(methods)
-import(xml2)
 importFrom(tools,file_ext)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
@@ -30,9 +29,20 @@ importFrom(xml2,as_list)
 importFrom(xml2,read_xml)
 importFrom(xml2,write_xml)
 importFrom(xml2,xml_add_child)
+importFrom(xml2,xml_add_parent)
+importFrom(xml2,xml_attr)
+importFrom(xml2,xml_attrs)
 importFrom(xml2,xml_children)
 importFrom(xml2,xml_dtd)
+importFrom(xml2,xml_find_all)
+importFrom(xml2,xml_name)
 importFrom(xml2,xml_new_root)
 importFrom(xml2,xml_ns)
+importFrom(xml2,xml_remove)
+importFrom(xml2,xml_root)
+importFrom(xml2,xml_set_attr)
+importFrom(xml2,xml_set_name)
 importFrom(xml2,xml_set_namespace)
+importFrom(xml2,xml_set_text)
+importFrom(xml2,xml_text)
 importFrom(xml2,xml_validate)
