### 2017-08-16 78a04c1
makes eml_view work (#223)
* makes eml_view work

* oops
