### 2017-10-23 9486f17
Add support for converting EML into JSON-LD (via Scema.org/Dataset)
This isn't feature complete but covers the basic stuff for now.
