+++ R/eml_to_jsonld.R
@@ -1,0 +1,317 @@
+#' Convert an EML abstract into its JSON-LD equivalent
+#'
+#' This function is convoluted. It takes an EML abstract, converts it into a
+#' DocBook XML document, uses Pandoc (via the rmarkdown package) to convert that
+#' into plain text, cleans up the output, and returns a character vector of
+#' length one.
+#'
+#' @param abstract
+#'
+#' @return (character) The abstract text as a single-length character vector
+eml_abstract_to_jsonld <- function(abstract) {
+  if (!requireNamespace("rmarkdown", quietly = TRUE)) {
+    stop(paste("The rmarkdown package is required to convert EML abstracts",
+               "into a format appropriate for Schema.org's description field"),
+         call. = FALSE)
+  }
+
+  if (!inherits(abstract, "abstract")) {
+    stop(call. = FALSE,
+         paste0("Argument 'abstract' must be of class 'abstract' but was of ",
+                "class '", class(abstract), "' instead."))
+  }
+
+  abstract_children <- xml2::xml_children(EML:::s4_to_xml(abstract))
+
+  # Create a bare-bones DocBook XML document to put the EML abstract's children
+  # into
+  doctype <-
+    xml2::xml_new_root(
+      xml2::xml_dtd("section",
+                    "-//OASIS//DTD DocBook XML V4.2//EN",
+                    "http://oasis-open.org/docbook/xml/4.5/docbookx.dtd"))
+  article_el <- xml2::xml_add_child(doctype, "article")
+
+  # Because EML abstract can have multiple top-level children, and xml2::add_child
+  # doesn't support adding a nodeset (i.e. multiple children at a time), we have
+  # to one each top-level child at time
+  lapply(abstract_children, function(node) { xml2::xml_add_child(article_el, node)})
+
+  input_path <- tempfile()
+  output_path <- tempfile()
+
+  xml2::write_xml(doctype, input_path)
+
+  pandoc_convert(
+    input_path,
+    to = "plain",
+    from = "docbook",
+    output = output_path,
+    options = "--wrap=none" # Default is to hard-wrap at 72 char columns
+  )
+
+  # Clean up
+  unlink(xml_path)
+  unlink(output_path)
+
+  # Return cleaned up version
+  paste(remove_extra_empty_lines(plain_abstract), collapse = "\n")
+}
+
+
+#' Remove leading, duplicate interior, and trailing empty strings
+#'
+#' This function is a workaround for the odd behavior Pandoc exhibits when
+#' converting from DocBook to plain (text) where leading, duplicate interior,
+#' and trailing empty lines are included in the output.
+#'
+#' @param lines (character) A character vector of the lines of text to modify
+#'
+#' @return (character)
+remove_extra_empty_lines <- function(lines) {
+  total_lines <- length(lines)
+  first_line <- FALSE # Whether we've iterated to the first non-zero-length line
+
+  for (line_no in seq_along(lines)) {
+    line_nchar <- nchar(lines[line_no])
+
+    # Remove leading empty lines
+    if (!first_line) {
+      if (line_nchar == 0) {
+        lines[line_no] <- NA
+      } else {
+        first_line <- TRUE
+      }
+    }
+
+    # Remove the current line if the current one is blank and we're not at last
+    # line end and the next line is blank too
+    if (line_nchar == 0 &&
+        total_lines > line_no &&
+        nchar(lines[line_no + 1]) == 0) {
+      lines[line_no] <- NA
+    }
+
+    # Remove the current line if it's the last one and it's blank
+    if (line_no == total_lines &&
+        line_nchar == 0) {
+      lines[line_no] <- NA
+    }
+  }
+
[…]
