+++ man/eml_party_to_jsonld.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_to_jsonld.R
+\name{eml_party_to_jsonld}
+\alias{eml_party_to_jsonld}
+\title{Convert an EML Party to a list suitable for conversion to Schema.org JSON-LD}
+\usage{
+eml_party_to_jsonld(party)
+}
+\arguments{
+\item{party}{Any EML party (e.g., creator, contact, etc.)}
+}
+\value{
+(list) A list of statements for insertion into the JSON-LD document
+}
+\description{
+This is a helper function to handle converting an EML party to a list with a
+structure suitable for conversion to JSON-LD. The main trick is that the type
+of an EML party is implicit based upon the presence of individualName,
+organizationName, and positionName children elements whereas the type in
+Schema.org JSON-LD is explicit.
+}
+\details{
+Also because Schema.org doesn't have a concept for a party that is defined
+only by its position, this function produces a warning and returns an empty
+list whenever the EML party is only defined by a position.
+}
