+++ man/eml_geographicCoverage_to_Place.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_to_jsonld.R
+\name{eml_geographicCoverage_to_Place}
+\alias{eml_geographicCoverage_to_Place}
+\title{Convert an EML geographiCoverage to a list suitable for conversion to
+Schema.org JSON-LD}
+\usage{
+eml_geographicCoverage_to_Place(geographicCoverage)
+}
+\arguments{
+\item{geographicCoverage}{An EML geographicCoverage object}
+}
+\value{
+(list) A list of statements for insertion into the JSON-LD document
+}
+\description{
+Convert an EML geographiCoverage to a list suitable for conversion to
+Schema.org JSON-LD
+}
