+++ man/eml_SingleDateTime_to_Text.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_to_jsonld.R
+\name{eml_SingleDateTime_to_Text}
+\alias{eml_SingleDateTime_to_Text}
+\title{Convert an EML SingleDateTime to a list suitable for conversion to Schema.org
+ JSON-LD}
+\usage{
+eml_SingleDateTime_to_Text(SingleDateTime)
+}
+\arguments{
+\item{SingleDateTime}{An EML SingleDateTime object}
+}
+\value{
+(character) The datetime as a string
+}
+\description{
+Convert an EML SingleDateTime to a list suitable for conversion to Schema.org
+ JSON-LD
+}
