+++ man/eml_to_jsonld.Rd
@@ -1,0 +1,42 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_to_jsonld.R
+\name{eml_to_jsonld}
+\alias{eml_to_jsonld}
+\title{Convert an EML object to a list suitable for conversion to Schema.org JSON-LD}
+\usage{
+eml_to_jsonld(doc)
+}
+\arguments{
+\item{doc}{An EML Object to convert}
+}
+\value{
+A JSON(-LD) string
+}
+\description{
+This function makes heavy use of streamlining assumptions about how to
+convert from EML to Schema.org JSON-LD. In-line comments describe the nuances
+of the conversion for each Schema.org concept.
+}
+\examples{
+library(EML)
+doc <- read_eml(system.file("examples", "hf205.xml", package = "EML"))
+eml_to_jsonld(doc)
+eml:abstract <-> schema:description
+
+This is done in sort of a roundabout fashion: Abstracts are converted
+from their EML DocBook-subset form by conveting them by hand into DocBook
+documents, running them thorugh Pandoc (via the rmarkdown package) and
+converting them to 'plain' format and cleaning up the output.
+creator
+TODO: Improve how much of this mapping is covered
+contact
+TODO: How does this map to Schema.org/Dataset?
+
+eml:keywordSet <-> schema:keywords
+
+Note: Returns the unique list of all keywords in the dataset description
+eml:geographiCoverage <-> schema:spatialCoverage
+eml:temporalCoverage <-> schema:temporalCoverage
+
+Note: This is only converting to Text and not DateTime
+}
