+++ man/eml_abstract_to_jsonld.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eml_to_jsonld.R
+\name{eml_abstract_to_jsonld}
+\alias{eml_abstract_to_jsonld}
+\title{Convert an EML abstract into its JSON-LD equivalent}
+\usage{
+eml_abstract_to_jsonld(abstract)
+}
+\arguments{
+\item{abstract}{}
+}
+\value{
+(character) The abstract text as a single-length character vector
+}
+\description{
+This function is convoluted. It takes an EML abstract, converts it into a
+DocBook XML document, uses Pandoc (via the rmarkdown package) to convert that
+into plain text, cleans up the output, and returns a character vector of
+length one.
+}
