### 2017-10-31 a667348
Make get_TextType actually work
Closes #229
