### 2017-10-31 72a3a53
Fix a bug where get TextType wasn't working for simple cases
Closes #230

Also adds a regression test
