### 2017-10-31 f1609b8
Normalize whitespace in get_TextType
