### 2017-10-31 cdf59b5
Fix typo in roxygen docs for get_TextType
