### 2017-10-31 2740b2e
Add news entry about recent fixes to get_TextType
