### 2017-10-31 c175d21
Add two more news.md notes
