+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `set_taxonomicCoverage` now supports expanding the provided species names into full taxonomic classifications via the taxize package
 * Fixed two bugs in `get_TextType` which were preventing the method from working under multiple scenarios
+* Improved `eml_view` so it works when there are XML comments in the EML record being viewed
+* Added the ability to `set_physical` to calculate file size and checksums automatically
 
 # EML 1.0.3
 
