### 2017-10-31 8bf05b7
Fix up get_TextType (#231)
* Make get_TextType actually work

Closes #229

* Fix a bug where get TextType wasn't working for simple cases

Closes #230

Also adds a regression test

* Normalize whitespace in get_TextType

* Fix typo in roxygen docs for get_TextType

* Add news entry about recent fixes to get_TextType

* Add two more news.md notes
