### 2017-12-07 3629365
Add an initial set of party and related helper functions
