+++ man/set_creator.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/set_party.R
+\name{set_creator}
+\alias{set_creator}
+\title{Create a creator object}
+\usage{
+set_creator(givenNames, surName, organizationNames, positionNames, addresses,
+  phones, electronicMailAddresses, onlineUrls, userIds, salutations)
+}
+\arguments{
+\item{givenNames}{(character) One or more given names for the party's individualName}
+
+\item{surName}{(character) The sur name of the party's individualName}
+
+\item{organizationNames}{(character) One or more organization names for the party}
+
+\item{positionNames}{(character) One or more position names for the party}
+
+\item{addresses}{(address) One or more addresses for the party}
+
+\item{phones}{(character|phone) One or more phones for the party}
+
+\item{electronicMailAddresses}{(character) One or more email addresses for the party}
+
+\item{onlineUrls}{(character) One or more online URLs for the party}
+
+\item{userIds}{(userId) One or more user IDs for the party}
+
+\item{salutations}{(character) One or more salutations for the party's individualName}
+}
+\value{
+(creator) A creator object
+}
+\description{
+Create a creator object
+}
+\examples{
+set_creator("Mary", "Scientist")
+}
