+++ man/set_phone.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/set_party.R
+\name{set_phone}
+\alias{set_phone}
+\title{A few helper methods to speed up creating various types of EML party
+objects (creator, contact, etc.).}
+\usage{
+set_phone(number, phonetype)
+}
+\arguments{
+\item{number}{(character) The phone number}
+
+\item{phonetype}{(character) The type of phone number}
+}
+\value{
+(phone) A phone object
+}
+\description{
+High-level functions include:
+}
+\details{
+- set_creator
+- set_contact
+- set_associatedParty
+- set_metadataProvider
+
+Each of these functions calls eml_party with the provided arguments.
+
+Extra helper functions included are:
+
+set_phone
+set_address
+set_userId
+Create a phone object
+}
+\examples{
+set_phone("1-555-555-5555")
+set_phone("1-555-555-5556", "facsimilie")
+}

[…]
