### 2017-12-07 db09d72
Fix various issues with last commit's args/docs
Made a few silly mistakes prior to committing.
1: failed to set custom args on all the functions I intended to
2: didn't scan all the generated roxygen docs for correctness
