### 2017-12-07 b75810b
Commit an updated Rd page for set_TextType
