### 2017-12-07 68df893
Fill in examples for new party helpers
