+++ man/set_creator.Rd
@@ -35,5 +35,21 @@ set_creator(givenNames, surName, organizationNames, positionNames, addresses,
 Create a creator object
 }
 \examples{
-set_creator("Mary", "Scientist")
+# A simple example
+set_creator("Bryce", "Mecum")
+# A more complete example
+set_creator("Bryce",
+            "Mecum",
+            "National Center for Ecological Analysis and Synthesis",
+            "Scientific Software Engineer",
+            set_address(c("735 State St.", "Suite 300"),
+                        "Santa Barbara",
+                        "California",
+                        "93101-5504",
+                        "USA"),
+            set_phone("1-555-555-5555"),
+            "mecum@nceas.ucsb.edu",
+            "https://brycemecum.com",
+            set_userId("https://orcid.org/0000-0002-0381-3766",
+                       "https://orcid.org"))
 }
