+++ man/set_metadataProvider.Rd
@@ -35,5 +35,21 @@ set_metadataProvider(givenNames, surName, organizationNames, positionNames,
 Create a metadataProvider object
 }
 \examples{
-set_metadataProvider("Mary", "Scientist")
+# A simple example
+set_metadataProvider("Bryce", "Mecum")
+# A more complete example
+set_metadataProvider("Bryce",
+                     "Mecum",
+                     "National Center for Ecological Analysis and Synthesis",
+                     "Scientific Software Engineer",
+                     set_address(c("735 State St.", "Suite 300"),
+                                 "Santa Barbara",
+                                 "California",
+                                 "93101-5504",
+                                 "USA"),
+                     set_phone("1-555-555-5555"),
+                     "mecum@nceas.ucsb.edu",
+                     "https://brycemecum.com",
+                     set_userId("https://orcid.org/0000-0002-0381-3766",
+                                "https://orcid.org"))
 }
