+++ man/set_contact.Rd
@@ -35,5 +35,21 @@ set_contact(givenNames, surName, organizationNames, positionNames, addresses,
 Create a contact object
 }
 \examples{
-set_contact("Mary", "Scientist")
+# A simple example
+set_contact("Bryce", "Mecum")
+# A more complete example
+set_contact("Bryce",
+            "Mecum",
+            "National Center for Ecological Analysis and Synthesis",
+            "Scientific Software Engineer",
+            set_address(c("735 State St.", "Suite 300"),
+                        "Santa Barbara",
+                        "California",
+                        "93101-5504",
+                        "USA"),
+            set_phone("1-555-555-5555"),
+            "mecum@nceas.ucsb.edu",
+            "https://brycemecum.com",
+            set_userId("https://orcid.org/0000-0002-0381-3766",
+                       "https://orcid.org"))
 }
