### 2017-12-07 6f0746c
Fix r cmd check issues with set_party
