+++ R/set_party.R
@@ -105,26 +105,6 @@ set_userId <- function(userId, directory) {
 #' @param role (character) A role for the party
 #'
 #' @return An object of the class equal to \code{type}
-#'
-#' @examples
-#' # A simple example
-#' set_party("metadata_provider", "Bryce", "Mecum")
-#' # A more complete example
-#' set_party("creator",
-#'           "Bryce",
-#'           "Mecum",
-#'           "National Center for Ecological Analysis and Synthesis",
-#'           "Scientific Software Engineer",
-#'           set_address(c("735 State St.", "Suite 300"),
-#'                       "Santa Barbara",
-#'                       "California",
-#'                       "93101-5504",
-#'                       "USA"),
-#'           set_phone("1-555-555-5555"),
-#'           "mecum@nceas.ucsb.edu",
-#'           "https://brycemecum.com",
-#'           set_userId("https://orcid.org/0000-0002-0381-3766",
-#'                      "https://orcid.org"))
 set_party <- function(type="associatedParty",
                       givenNames,
                       surName,
@@ -145,7 +125,7 @@ set_party <- function(type="associatedParty",
                 "' and the role argument is only appropriate when type is assoicatedParty."))
   }
 
-  #' Stop if no individualName, organizationName, or positionName are set
+  # Stop if no individualName, organizationName, or positionName are set
   if (missing(surName) && missing(organizationNames) && missing(positionNames)) {
     stop("At least one of surName, organizationNames, or positionNames must be set.")
   }
