### 2017-12-07 3a53724
Add some basic unit and regression tests for set_party
