### 2017-12-07 597a114
Add a few more regression tests
