### 2017-12-07 b8e3fbb
Add an initial set of party and related helper functions (#232)
* Add an initial set of party and related helper functions

* Fix various issues with last commit's args/docs

Made a few silly mistakes prior to committing.
1: failed to set custom args on all the functions I intended to
2: didn't scan all the generated roxygen docs for correctness

* Fill in examples for new party helpers

* Fix r cmd check issues with set_party

* Add some basic unit and regression tests for set_party

* Add a few more regression tests
