+++ R/set_party.R
@@ -1,0 +1,435 @@
+#' Create a phone object
+#'
+#' @param number (character) The phone number
+#' @param phonetype (character) The type of phone number
+#'
+#' @return (phone) A phone object
+#' @export
+#'
+#' @examples
+#' set_phone("1-555-555-5555")
+#' set_phone("1-555-555-5556", "facsimilie")
+set_phone <- function(number, phonetype) {
+  phone <- new("phone",
+               .Data = number)
+
+  if (!missing(phonetype)) {
+    phone@phonetype <- as(phonetype, "xml_attribute")
+  }
+
+  phone
+}
+
+#' Create an \code{address} object
+#'
+#' @param deliveryPoints (character|deliveryPoint) The delivery point (i.e., street address)
+#' @param city (character|city) The city
+#' @param administrativeArea (character|administrativeArea) The administrative area (i.e., the state (in the USA at least))
+#' @param postalCode (character|postalCode) The postal code (AKA zip code (in the USA))
+#' @param country (character|country) The country
+#'
+#' @return (address) An address object
+#' @export
+#'
+#' @examples
+#' set_address("1 Infinite Loop", "Cupertino", "CA", "95014", "USA")
+#' set_address(administrativeArea = "California", country = "USA")
+set_address <- function(deliveryPoints, city, administrativeArea, postalCode, country) {
+  address <- new("address")
+
+  if (!missing(deliveryPoints)) {
+    address@deliveryPoint <- new("ListOfdeliveryPoint",
+                                 .Data = lapply(deliveryPoints, as, "deliveryPoint"))
+  }
+
+  if (!missing(city)) {
+    address@city <- as(city, "city")
+  }
+
+  if (!missing(administrativeArea)) {
+    address@administrativeArea <- as(administrativeArea, "administrativeArea")
+  }
+
+  if (!missing(postalCode)) {
+    address@postalCode <- as(postalCode, "postalCode")
+  }
+
+  if (!missing(country)) {
+    address@country <- as(country, "country")
+  }
+
+  address
+}
+
+#' Create a \code{userId} object
+#'
+#' @param userId (character) The user ID
+#' @param directory (character) The directory under which the user ID applies
+#'
+#' @return (userId) A userId object
+#' @export
+#'
+#' @examples
+#' set_userId("alice", "ldap://ldap.example.org")
+#' set_userId("https://orcid.org/0000-0002-0381-3766", "https://orcid.org")
+set_userId <- function(userId, directory) {
+  user_id <- new("userId")
+
+  user_id@.Data <- userId
+  user_id@directory <- as(directory, "xml_attribute")
+
+  user_id
+}
+
+#' Create a party object
+#'
+#' This function is meant to be used by other functions to create particular
+#' types of EML parties (e.g., creator)
+#'
+#' Arguments, in most cases, are coerced to their appropriate type so many
+#' of them can receive character vectors while they can also take vectors of
+#' the appropriate target type (though this is not explicitly laid out in the
+#' documentation).
+#'
+#' @param type (character) The type of party to create
+#' @param givenNames (character) One or more given names for the party's individualName
+#' @param surName (character) The sur name of the party's individualName
+#' @param organizationNames (character) One or more organization names for the party
+#' @param positionNames (character) One or more position names for the party
+#' @param addresses (address) One or more addresses for the party
+#' @param phones (character|phone) One or more phones for the party
+#' @param electronicMailAddresses (character) One or more email addresses for the party
+#' @param onlineUrls (character) One or more online URLs for the party
+#' @param userIds (userId) One or more user IDs for the party
+#' @param salutations (character) One or more salutations for the party's individualName
+#' @param role (character) A role for the party
+#'
+#' @return An object of the class equal to \code{type}
+set_party <- function(type="associatedParty",
+                      givenNames,
[…]
