+++ man/set_address.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/set_party.R
+\name{set_address}
+\alias{set_address}
+\title{Create an \code{address} object}
+\usage{
+set_address(deliveryPoints, city, administrativeArea, postalCode, country)
+}
+\arguments{
+\item{deliveryPoints}{(character|deliveryPoint) The delivery point (i.e., street address)}
+
+\item{city}{(character|city) The city}
+
+\item{administrativeArea}{(character|administrativeArea) The administrative area (i.e., the state (in the USA at least))}
+
+\item{postalCode}{(character|postalCode) The postal code (AKA zip code (in the USA))}
+
+\item{country}{(character|country) The country}
+}
+\value{
+(address) An address object
+}
+\description{
+Create an \code{address} object
+}
+\examples{
+set_address("1 Infinite Loop", "Cupertino", "CA", "95014", "USA")
+set_address(administrativeArea = "California", country = "USA")
+}
