### 2018-02-09 2d38f4d
Update Rd for set_physical
