### 2018-02-09 3bed6b2
Fix mix-up of LFCR -> CRLF as default recordDelimiter in set_physical. (#235)
* Fix mix-up of LFCR -> CRLF as default recordDelimiter in set_physical.

* Update Rd for set_physical

* Add note in NEWS about set_physical fix
