+++ R/set_physical.R
@@ -32,7 +32,7 @@ set_physical <- function(objectName,
                          id = character(),
                          numHeaderLines = character(),
                          numFooterLines = character(),
-                         recordDelimiter = "\\n\\r",
+                         recordDelimiter = "\\r\\n",
                          fieldDelimiter = ",",
                          collapseDelimiters = logical(),
                          literalCharacter = character(),
