### 2018-02-22 c45674e
updated get_attributes to exclude merging NULL data frames
