### 2018-02-22 760dcb1
reduced two helper functions to one
