### 2018-02-23 679c126
get attributes bug fix for NULL data frames (#238)
* updated get_attributes to exclude merging NULL data frames

* reduced two helper functions to one

* added unit tests for get_attributes

* fix merge_df bug
