### 2018-04-04 63dfd29
drop devel checks on travis for now
apparently travis has no idea how to set itself up for R devel right now.... fails to install devtools during the travis R setup stage.
