### 2018-04-05 f5fdf7a
Fix small typo (#241)
